docs: aria menubar examples
