+++ adev/src/content/examples/aria/menu/src/menu-trigger/material/app/app.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/icon?family=Material+Symbols+Outlined');
+
 :host {
   display: flex;
   justify-content: center;
