+++ adev/src/content/examples/aria/menubar/src/rtl/app/app.ts
@@ -1,8 +1,34 @@
-import {Component} from '@angular/core';
+import {AfterViewInit, Component, signal, viewChild} from '@angular/core';
+import {MenuBar, Menu, MenuContent, MenuItem} from '@angular/aria/menu';
+import {OverlayModule} from '@angular/cdk/overlay';
+import {Dir} from '@angular/cdk/bidi';
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.html',
-  styleUrl: './app.css',
+  templateUrl: 'app.html',
+  styleUrl: 'app.css',
+  imports: [Dir, MenuBar, Menu, MenuContent, MenuItem, OverlayModule],
+  standalone: true,
 })
-export class App {}
+export class App implements AfterViewInit {
+  fileMenu = viewChild<Menu<string>>('fileMenu');
+  shareMenu = viewChild<Menu<string>>('shareMenu');
+  editMenu = viewChild<Menu<string>>('editMenu');
+  viewMenu = viewChild<Menu<string>>('viewMenu');
+  insertMenu = viewChild<Menu<string>>('insertMenu');
+  imageMenu = viewChild<Menu<string>>('imageMenu');
+  chartMenu = viewChild<Menu<string>>('chartMenu');
+  formatMenu = viewChild<Menu<string>>('formatMenu');
+  textMenu = viewChild<Menu<string>>('textMenu');
+  sizeMenu = viewChild<Menu<string>>('sizeMenu');
+  paragraphMenu = viewChild<Menu<string>>('paragraphMenu');
+  alignMenu = viewChild<Menu<string>>('alignMenu');
+
+  rendered = signal(false);
+
+  ngAfterViewInit() {
+    setTimeout(() => {
+      debugger;
+    }, 500);
+  }
+}
