+++ adev/src/content/examples/aria/menubar/src/rtl/material/app/app.ts
@@ -1,15 +1,16 @@
-import {AfterViewInit, Component, signal, viewChild} from '@angular/core';
+import {Component, viewChild} from '@angular/core';
 import {MenuBar, Menu, MenuContent, MenuItem} from '@angular/aria/menu';
 import {OverlayModule} from '@angular/cdk/overlay';
+import {Dir} from '@angular/cdk/bidi';
 
 @Component({
   selector: 'app-root',
   templateUrl: 'app.html',
   styleUrl: 'app.css',
-  imports: [MenuBar, Menu, MenuContent, MenuItem, OverlayModule],
+  imports: [Dir, MenuBar, Menu, MenuContent, MenuItem, OverlayModule],
   standalone: true,
 })
-export class App implements AfterViewInit {
+export class App {
   fileMenu = viewChild<Menu<string>>('fileMenu');
   shareMenu = viewChild<Menu<string>>('shareMenu');
   editMenu = viewChild<Menu<string>>('editMenu');
@@ -22,12 +23,4 @@ export class App implements AfterViewInit {
   sizeMenu = viewChild<Menu<string>>('sizeMenu');
   paragraphMenu = viewChild<Menu<string>>('paragraphMenu');
   alignMenu = viewChild<Menu<string>>('alignMenu');
-
-  rendered = signal(false);
-
-  ngAfterViewInit() {
-    setTimeout(() => {
-      debugger;
-    }, 500);
-  }
 }
